+++ src/add-ons/kernel/file_systems/ext2/HTreeEntryIterator.h
@@ -34,8 +34,8 @@ public:
 
 			uint32				BlocksNeededForNewEntry();
 			status_t			InsertEntry(Transaction& transaction,
-									uint32 hash, uint32 block,
-									uint32 newBlocksPos, bool hasCollision);
+									uint32 hash, off_t block,
+									off_t newBlocksPos, bool hasCollision);
 private:
 								HTreeEntryIterator(uint32 block,
 									uint32 blockSize, Inode* directory,
@@ -53,7 +53,7 @@ private:
 			uint16				fCurrentEntry;
 
 			uint32				fBlockSize;
-			uint32				fBlockNum;
+			off_t				fBlockNum;
 
 			HTreeEntryIterator*	fParent;
 			HTreeEntryIterator*	fChild;
